requirements.txt
fastapi
uvicorn
groq
python-dotenv
requests
pdfminer.six
streamlit
python-multipart